Revise package structure
