+++ plugin/plugin.go
@@ -8,30 +8,28 @@ import (
 	"github.com/terraform-linters/tflint-plugin-sdk/tflint"
 )
 
-// handShakeConfig is used for UX. ProcotolVersion will be updated by incompatible changes
+// handShakeConfig is used for UX. ProcotolVersion will be updated by incompatible changes.
 var handshakeConfig = plugin.HandshakeConfig{
 	ProtocolVersion:  1,
 	MagicCookieKey:   "TFLINT_RULESET_PLUGIN",
 	MagicCookieValue: "5adSn1bX8nrDfgBqiAqqEkC6OE1h3iD8SqbMc5UUONx8x3xCF0KlPDsBRNDjoYDP",
 }
 
-// RuleSetPlugin is a wrapper to satisfy the interface of go-plugin
+// RuleSetPlugin is a wrapper to satisfy the interface of go-plugin.
 type RuleSetPlugin struct {
 	impl tflint.RuleSet
 }
 
-// Server returns an RPC server acting as a plugin
+// Server returns an RPC server acting as a plugin.
 func (p *RuleSetPlugin) Server(b *plugin.MuxBroker) (interface{}, error) {
 	return &Server{impl: p.impl, broker: b}, nil
 }
 
-// Client returns an RPC client for use by the host
+// Client returns an RPC client for the host.
 func (RuleSetPlugin) Client(b *plugin.MuxBroker, c *rpc.Client) (interface{}, error) {
 	return &Client{rpcClient: c, broker: b}, nil
 }
 
-// In order to communicate the interface correctly with RPC,
-// the type of the related structure is registered in gob at the initial time.
 func init() {
 	gob.Register(tflint.Error{})
 }
