+++ plugin/server.go
@@ -3,21 +3,22 @@ package plugin
 import (
 	plugin "github.com/hashicorp/go-plugin"
 	"github.com/terraform-linters/tflint-plugin-sdk/tflint"
+	tfclient "github.com/terraform-linters/tflint-plugin-sdk/tflint/client"
 )
 
-// Server is an RPC server acting as a plugin
+// Server is an RPC server acting as a plugin.
 type Server struct {
 	impl   tflint.RuleSet
 	broker *plugin.MuxBroker
 }
 
-// ServeOpts is an option for serving a plugin
-// Each plugin can pass a RuleSet that represents its own functionality
+// ServeOpts is an option for serving a plugin.
+// Each plugin can pass a RuleSet that represents its own functionality.
 type ServeOpts struct {
 	RuleSet tflint.RuleSet
 }
 
-// Serve is a wrapper of plugin.Serve. This is entrypoint of all plugins
+// Serve is a wrapper of plugin.Serve. This is entrypoint of all plugins.
 func Serve(opts *ServeOpts) {
 	plugin.Serve(&plugin.ServeConfig{
 		HandshakeConfig: handshakeConfig,
@@ -27,36 +28,37 @@ func Serve(opts *ServeOpts) {
 	})
 }
 
-// RuleSetName replies its own the result of RuleSetName
+// RuleSetName returns the name of the plugin.
 func (s *Server) RuleSetName(args interface{}, resp *string) error {
 	*resp = s.impl.RuleSetName()
 	return nil
 }
 
-// RuleSetVersion replies its own the result of RuleSetVersion
+// RuleSetVersion returns the version of the plugin.
 func (s *Server) RuleSetVersion(args interface{}, resp *string) error {
 	*resp = s.impl.RuleSetVersion()
 	return nil
 }
 
-// RuleNames replies its own the result of RuleNames
+// RuleNames returns the list of rule names provided by the plugin.
 func (s *Server) RuleNames(args interface{}, resp *[]string) error {
 	*resp = s.impl.RuleNames()
 	return nil
 }
 
-// ApplyConfig applies the passed config to its own plugin implementation
+// ApplyConfig applies the passed config to its own plugin implementation.
 func (s *Server) ApplyConfig(config *tflint.Config, resp *interface{}) error {
 	s.impl.ApplyConfig(config)
 	return nil
 }
 
-// Check initializes an RPC client that can query to the host process and pass it to the Check method
+// Check calls its own plugin implementation with an RPC client that can send
+// requests to the host process.
 func (s *Server) Check(brokerID uint32, resp *interface{}) error {
 	conn, err := s.broker.Dial(brokerID)
 	if err != nil {
 		return err
 	}
 
-	return s.impl.Check(tflint.NewClient(conn))
+	return s.impl.Check(tfclient.NewClient(conn))
 }
