+++ tflint/interface.go
@@ -7,26 +7,48 @@ import (
 
 // Runner acts as a client for each plugin to query the host process about the Terraform configurations.
 type Runner interface {
+	// WalkResourceAttributes visits attributes with the passed function.
+	// You must pass a resource type as the first argument and an attribute name as the second argument.
 	WalkResourceAttributes(string, string, func(*hcl.Attribute) error) error
+
+	// WalkResourceBlocks visits blocks with the passed function.
+	// You must pass a resource type as the first argument and a block type as the second argument.
+	// This API currently does not support labeled blocks.
 	WalkResourceBlocks(string, string, func(*hcl.Block) error) error
+
+	// WalkResources visits resources with the passed function.
+	// You must pass a resource type as the first argument.
 	WalkResources(string, func(*terraform.Resource) error) error
+
+	// EvaluateExpr evaluates the passed expression and reflects the result in ret.
+	// Since this function returns an application error, it is expected to use the EnsureNoError
+	// to determine whether to continue processing.
 	EvaluateExpr(expr hcl.Expression, ret interface{}) error
+
+	// EmitIssue sends an issue to TFLint. You need to pass the message of the issue and the range.
 	EmitIssue(rule Rule, message string, location hcl.Range, meta Metadata) error
+
+	// EnsureNoError is a helper for error handling. Depending on the type of error generated by EvaluateExpr,
+	// determine whether to exit, skip, or continue. If it is continued, the passed function will be executed.
 	EnsureNoError(error, func() error) error
 }
 
 // Rule is the interface that the plugin's rules should satisfy.
 type Rule interface {
+	// Name will be displayed with a message of an issue and will be the identifier used to control
+	// the behavior of this rule in the configuration file etc.
+	// Therefore, it is expected that this will not duplicate the rule names provided by other plugins.
 	Name() string
-	Enabled() bool
-	Severity() string
-	Link() string
-	Check(Runner) error
-}
 
-// Server is the interface that hosts that provide the plugin mechanism must meet in order to respond to queries from the plugin.
-type Server interface {
-	Attributes(*AttributesRequest, *AttributesResponse) error
-	EvalExpr(*EvalExprRequest, *EvalExprResponse) error
-	EmitIssue(*EmitIssueRequest, *interface{}) error
+	// Enabled indicates whether the rule is enabled by default.
+	Enabled() bool
+
+	// Severity indicates the severity of the rule.
+	Severity() string
+
+	// Link allows you to add a reference link to the rule.
+	Link() string
+
+	// Check is the entrypoint of the rule. You can fetch Terraform configurations and send issues via Runner.
+	Check(Runner) error
 }
