+++ tflint/issue.go
@@ -2,6 +2,7 @@ package tflint
 
 import hcl "github.com/hashicorp/hcl/v2"
 
+// List of issue severity levels. The rules implemented by a plugin can be set to any severity.
 const (
 	// ERROR is possible errors
 	ERROR = "Error"
@@ -11,43 +12,7 @@ const (
 	NOTICE = "Notice"
 )
 
-// Metadata is the additional data sent to the host process to build the issue.
+// Metadata is an additional data sent to the host process to build the issue.
 type Metadata struct {
 	Expr hcl.Expression
 }
-
-// RuleObject is an intermediate representation for communicating with RPC.
-type RuleObject struct {
-	Data *RuleObjectData
-}
-
-// RuleObjectData holds the data that RuleObject needs to satisfy the Rule interface.
-type RuleObjectData struct {
-	Name     string
-	Enabled  bool
-	Severity string
-	Link     string
-}
-
-func newObjectFromRule(rule Rule) *RuleObject {
-	return &RuleObject{
-		Data: &RuleObjectData{
-			Name:     rule.Name(),
-			Enabled:  rule.Enabled(),
-			Severity: rule.Severity(),
-			Link:     rule.Link(),
-		},
-	}
-}
-
-// Name is a reference method to internal data
-func (r *RuleObject) Name() string { return r.Data.Name }
-
-// Enabled is a reference method to internal data
-func (r *RuleObject) Enabled() bool { return r.Data.Enabled }
-
-// Severity is a reference method to internal data
-func (r *RuleObject) Severity() string { return r.Data.Severity }
-
-// Link is a reference method to internal data
-func (r *RuleObject) Link() string { return r.Data.Link }
